+++ NINA.Plugin.Web/NINA.Plugin.Web/Properties/Settings.Designer.cs
@@ -12,7 +12,7 @@ namespace Web.NINAPlugin.Properties {
     
     
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "17.0.3.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "17.1.0.0")]
     internal sealed partial class Settings : global::System.Configuration.ApplicationSettingsBase {
         
         private static Settings defaultInstance = ((Settings)(global::System.Configuration.ApplicationSettingsBase.Synchronized(new Settings())));
@@ -32,30 +32,6 @@ namespace Web.NINAPlugin.Properties {
             }
             set {
                 this["UpdateSettings"] = value;
-            }
-        }
-        
-        [global::System.Configuration.UserScopedSettingAttribute()]
-        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("False")]
-        public bool WebPluginEnabled {
-            get {
-                return ((bool)(this["WebPluginEnabled"]));
-            }
-            set {
-                this["WebPluginEnabled"] = value;
-            }
-        }
-        
-        [global::System.Configuration.UserScopedSettingAttribute()]
-        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("10")]
-        public int PurgeDays {
-            get {
-                return ((int)(this["PurgeDays"]));
-            }
-            set {
-                this["PurgeDays"] = value;
             }
         }
         
@@ -85,18 +61,6 @@ namespace Web.NINAPlugin.Properties {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("80")]
-        public int WebServerPort {
-            get {
-                return ((int)(this["WebServerPort"]));
-            }
-            set {
-                this["WebServerPort"] = value;
-            }
-        }
-        
-        [global::System.Configuration.UserScopedSettingAttribute()]
-        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
         [global::System.Configuration.DefaultSettingValueAttribute("")]
         public string HostAddress {
             get {
@@ -104,18 +68,6 @@ namespace Web.NINAPlugin.Properties {
             }
             set {
                 this["HostAddress"] = value;
-            }
-        }
-        
-        [global::System.Configuration.UserScopedSettingAttribute()]
-        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("False")]
-        public bool NonLights {
-            get {
-                return ((bool)(this["NonLights"]));
-            }
-            set {
-                this["NonLights"] = value;
             }
         }
     }